+++ components/index.ts
@@ -4,3 +4,4 @@ export { BonkSlider } from './BonkSlider';
 export { Scanner } from './Scanner';
 export { DiscoverFeed } from './DiscoverFeed';
 export { StatusBar } from './StatusBar';
+export { ImageViewer } from './ImageViewer';